Change Plane intersect to return pixel coordinates with boundary checking
- Modified Plane::intersect to return Option<(usize, usize)> instead of Option<(f64, f64)>
- Added boundary checking to return None if intersection is outside plane dimensions
- Plane coordinate is treated as center, with width/height defining the extent
- Updated Projection::project return type to match new intersect signature
- Simplified map_to_pixels2d to use pixel coordinates directly
- Removed HALF_SIZE constant as coordinates are now in pixel space

This change makes the coordinate system more intuitive by returning pixel
coordinates directly and enforcing plane boundaries at the intersection level.

+++ src/camera.rs
@@ -14,7 +14,7 @@ pub struct Projection {
 }
 
 impl Projection {
-    pub fn project(&self, target: &Tuple3D) -> Option<(f64, (f64, f64))> {
+    pub fn project(&self, target: &Tuple3D) -> Option<(f64, (usize, usize))> {
         let relative_option = self.screen.intersect(&self.camera, target);
         if let Some(relative_coords) = relative_option {
             let distance = self.camera.coordinate_squared_distance(target);
